requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0
tqdm>=4.65.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.0.280
jupyterlab>=4.0.0
scipy>=1.11.0
